+++ .gitignore
@@ -108,7 +108,6 @@ dist
 
 # Some commands hush hush (Most are WIP)
 sms.*
-events
 cmdsOLDOLD
 add*
 bal*
