+++ bot.js
@@ -117,7 +117,7 @@ for(const file of slashCommandFiles) {
     })
 	}
 }
-/* let evAlpha = {}
+let evAlpha = {}
 const eventFiles = fs.readdirSync('events/').filter(file => file.endsWith('.js'))
 console.log(chalk.magenta('[Karen Bot]'), chalk.yellow(`[Event]`), chalk.white('[Load]'), `Loading a total of ${eventFiles.length} events`)
 for(const ev of eventFiles) {
@@ -149,7 +149,7 @@ for(const ev of eventFiles) {
       }
     })
 	}
-} */
+}
 
 // Gets settingsmap
 axios({
@@ -182,261 +182,6 @@ axios({
 }).then(res => {
   guildProfile = new Map(res.data);
 })
-
-//* A bunch of functions for guild settings and stuff
-
-let statusQuotes = [
-  "5G causes corona cancer",
-  "I have a medical condition that doesn't let me wear masks",
-  "Corona vaccines put Microsoft chips in you",
-  "Vaccines cause autism",
-  "The earth is flat",
-  "One two three four you all suck and need to get baptised",
-  "im gay for jesus mom, mary",
-  "THE LOCHNESS MONSTER HAS RETURNED",
-  "AAAAAAAAAAAAAAAAAAAAAAA",
-  "THE PURGE IS COMING",
-  "WHERE ARE MY KIDS",
-  "luki works for the government",
-  "I ATE MARIJUANA BREAD PLEASE HELP",
-  "are you vaccinated? no? good",
-  "BUT THIS MOMMY BLOG SAYS ITS BAD",
-  "facebook is best social media"
-]
-
-client.once('ready', async () => {
-  logger.log('info', `Bot has started, with ${client.users.cache.size} users, in ${client.channels.cache.size} channels of ${client.guilds.cache.size} guilds.`);
-
-  if (process.env.VALIDATION  == undefined) {
-    axios({
-      "method": "POST",
-      "url": `${process.env.API_SERVER}/karen/logs/`,
-      "headers": {
-        "Authorization": process.env.AUTH_B64,
-        "Content-Type": "application/json; charset=utf-8",
-        'User-Agent': process.env.AUTH_USERAGENT
-      },
-      "auth": {
-        "username": process.env.AUTH_USER,
-        "password": process.env.AUTH_PASS
-      },
-      "data": {
-        "content": `Bot has started, with ${client.users.cache.size} users, in ${client.channels.cache.size} channels of ${client.guilds.cache.size} guilds.`,
-        "type": "info"
-      }
-    }) 
-  }
-
-  /* client.user.setActivity(`${client.guilds.cache.size} servers | ` + config.prefix +`help`, { type: "WATCHING" }); */
-  let why = statusQuotes[Math.floor(Math.random()*statusQuotes.length)];
-  // emergency status
-  //why = "⚠️ WELCOME FUNCTIONALITY DISABLED ⚠️"
-  client.user.setActivity(config.prefix +`help | ${why}`, { type: "WATCHING" });
-
-  let statsTimeout = () => {
-    const promises = [
-      client.shard.fetchClientValues('guilds.cache.size'),
-      client.shard.broadcastEval('this.guilds.cache.reduce((prev, guild) => prev + guild.memberCount, 0)'),
-    ];
-
-    Promise.all(promises)
-    .then(results => {
-      const totalGuilds = results[0].reduce((prev, guildCount) => prev + guildCount, 0);
-      const totalMembers = results[1].reduce((prev, memberCount) => prev + memberCount, 0);
-
-      axios({
-        "method": "POST",
-        "url": `${process.env.API_SERVER}/karen/stats/`,
-        "headers": {
-          "Authorization": process.env.AUTH_B64,
-          "Content-Type": "application/json; charset=utf-8",
-          'User-Agent': process.env.AUTH_USERAGENT
-        },
-        "auth": {
-          "username": process.env.AUTH_USER,
-          "password": process.env.AUTH_PASS
-        },
-        "data": {
-          "servercount": `${totalGuilds}`,
-          "users": `${totalMembers}`,
-          "DiscordJS": `${Discord.version}`
-        }
-      })
-    })
-    .catch(console.error);
-  }
-  if (process.env.VALIDATION  == undefined) {
-    statsTimeout()
-    var myVar = setInterval(statsTimeout, 600000)
-  }
-
-  // ------------- Slash commands -------------
-
-  // Puts the slash commands in ./slashcommands/ in an array
-  let slashCommandsArr = client.slashcommands.array()
-
-  // For each command, create a slash command
-  for(const command of slashCommandsArr) {
-    //                                            Only used for testing
-    client.api.applications(client.user.id)/* .guilds('701064832136249355') */.commands.post({data: {
-      "name": command.name,
-      "description": command.description,
-      "options": command.options
-    }})
-  }
-
-  // When a slash command is triggered
-  client.ws.on('INTERACTION_CREATE', async interaction => {
-
-    // Get the slash command that was triggered and  assign it to command
-    var slashCommandName = interaction.data.name
-    let command = client.slashcommands.get(slashCommandName)
-        || client.slashcommands.find(c => c.aliases && c.aliases.includes(slashCommandName))
-
-
-    // Execute command
-    try {
-      await command.execute(client, interaction)
-    } catch (error) {
-      console.error(error)
-      axios({
-        "method": "POST",
-        "url": `${process.env.API_SERVER}/karen/logs/`,
-        "headers": {
-          "Authorization": process.env.AUTH_B64,
-          "Content-Type": "application/json; charset=utf-8",
-          'User-Agent': process.env.AUTH_USERAGENT
-        },
-        "auth": {
-          "username": process.env.AUTH_USER,
-          "password": process.env.AUTH_PASS
-        },
-        "data": {
-          "content": `Slash command "${command.name}" by ${interaction.member.user.id} - ${error}`,
-          "type": "error"
-        }
-      })
-      axios.post(process.env.REGULAR_WEBHOOK, {
-        "content": `Slash command "${command.name}" by ${interaction.member.user.id} - ${error}`,
-        "embeds": null,
-        "username": 'Karen Bot Error (Slash)',
-        "avatar_url": "https://karen.exerra.xyz/assets/BotLogoNoOutline.png"
-      })
-    }
-  })
-});
-client.once('reconnecting', async () => {
-  axios.post(`${process.env.API_SERVER}/karen/logs/`, {
-    "content": `Reconnecting!`,
-    "type": 'info'
-  })
-  let why = statusQuotes[Math.floor(Math.random()*statusQuotes.length)];
-  // emergency status
-  //why = "⚠️ WELCOME FUNCTIONALITY DISABLED ⚠️"
-  client.user.setActivity(config.prefix +`help | ${why}`, { type: "WATCHING" });
-});
-client.once('disconnect', async () => {
-  axios.post(`${process.env.API_SERVER}/karen/logs/`, {
-    "content": `Disconnect!`,
-    "type": 'info'
-  })
-  let why = statusQuotes[Math.floor(Math.random()*statusQuotes.length)];
-  // emergency status
-  //why = "⚠️ WELCOME FUNCTIONALITY DISABLED ⚠️"
-  client.user.setActivity(config.prefix +`help | ${why}`, { type: "WATCHING" });
-});
-
-client.on("guildDelete", async guild => {
-  // this event triggers when the bot is removed from a guild.
-  axios.post(`${process.env.API_SERVER}/karen/logs/`, {
-    "content": `I have been removed from: ${guild.name} (id: ${guild.id})`,
-    "type": 'info'
-  })
-  let why = await (await fetch(`https://nekos.life/api/v2/why`)).json() // skipcq: JS-0128
-});
-
-client.on("guildCreate", async guild => {
-  // This event triggers when the bot joins a guild.
-  axios.post(`${process.env.API_SERVER}/karen/logs/`, {
-    "content": `New guild joined: ${guild.name} (id: ${guild.id}). This guild has ${guild.memberCount} members!`,
-    "type": 'info'
-  })
-  let why = await (await fetch(`https://nekos.life/api/v2/why`)).json() // skipcq: JS-0128
-});
-
-client.on('guildMemberAdd', async member => {
-  let why = await (await fetch(`https://nekos.life/api/v2/why`)).json() // skipcq: JS-0128
-  axios({
-    "method": "POST",
-    "url": `${process.env.API_SERVER}/karen/profile/get/`,
-    "headers": {
-      "Authorization": process.env.AUTH_B64,
-      "Content-Type": "application/json; charset=utf-8",
-      'User-Agent': process.env.AUTH_USERAGENT
-    },
-    "auth": {
-      "username": process.env.AUTH_USER,
-      "password": process.env.AUTH_PASS
-    },
-    "data": {
-      "id": member.id
-    }
-  }).then((response) => {
-    // If success, return
-    return
-  }, (error) => {
-    // If error (which means person doesn't have a profile), create one
-    const profile = {
-      description: "None",
-      gender: "Not specified",
-      birthday: "Not specified",
-      country: "None",
-      rank: "",
-      languages: "None"
-    }
-    // Sends profile to server
-    axios({
-      "method": "POST",
-      "url": `${process.env.API_SERVER}/karen/profile/`,
-      "headers": {
-        "Authorization": process.env.AUTH_B64,
-        "Content-Type": "application/json; charset=utf-8",
-        'User-Agent': process.env.AUTH_USERAGENT
-      },
-      "auth": {
-        "username": process.env.AUTH_USER,
-        "password": process.env.AUTH_PASS
-      },
-      "data": {
-        profile,
-        "id": member.id
-      }
-    })
-  });
-
-  // this is my shitty way of handling role adds
-  // dont call me out on this, this is our little secret ok?
-  if (member.guild.id == '826439565086425098') {
-    let role = member.guild.roles.cache.find(r => r.name === "peasants");
-    member.roles.add(role)
-  }
-  if (member.guild.id == '828999868239708220') {
-    let role1 = member.guild.roles.cache.get("829007236222812222");
-    let role2 = member.guild.roles.cache.get("829007055745974312");
-    let role3 = member.guild.roles.cache.get("829006864574185502");
-    let role4 = member.guild.roles.cache.get("829032788652195850");
-    let role5 = member.guild.roles.cache.get('829005429655863387')
-    member.roles.add(role1)
-    member.roles.add(role2)
-    member.roles.add(role3)
-    member.roles.add(role4)
-    member.roles.add(role5)
-  }
-  if (settingsmap.get(member.guild.id).welcomeEnabled == false) return;
-  const channel = member.guild.channels.cache.find(ch => ch.id == settingsmap.get(member.guild.id).welcomeChannel);
-  channel.send(`Welcome to this server, ${member}.`);
-  member.send('READ THE RULES\nDO IT NOW\nYOU BETTER FUCKING DO IT\n\n\n~~dont take it seriously, but like please do read the rules :)~~ ')
-});
 
 client.on('message', async msg => {
