+++ package.json
@@ -35,6 +35,7 @@
     "node-fetch": "^2.6.1",
     "node-spotify-api": "^1.1.1",
     "nsfai": "^2.0.0",
+    "ora": "^5.4.1",
     "qrcode": "^1.4.4",
     "turndown": "^5.0.1",
     "twitter": "^1.7.1",
